+++ src/index.html
@@ -11,7 +11,7 @@
     integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p"
     crossorigin="anonymous"
     /> 
-    <title>Wbpack Exercise</title>
+    <title>Todo List</title>
 </head>
 <body>
     <div id="to-do-list">
@@ -23,7 +23,7 @@
             <li>
                 <form id="form">
                     <input id="add" type="text" placeholder="Add to your list..." required> 
-                    <button id="addBtn" type="submit" ><i class="fa fa-share"></i></button>
+                    <button id="addBtn" type="button" ><i class="fa fa-share"></i></button>
 
                 </form>
