+++ src/index.js
@@ -1,8 +1,15 @@
 import './style.css';
-import {
-  checkbox, addBtn, input, clearCompleted, taskBox, deleteIcon, toHide,
-} from './variables.js';
 import ToDo from './classes.js';
+
+export const containerElement = document.getElementById('container');
+export const checkbox = document.getElementsByClassName('checkbox');
+
+const addBtn = document.querySelector('#addBtn');
+const input = document.querySelector('#add');
+const clearCompleted = document.querySelector('#clear-completed');
+const taskBox = document.getElementsByClassName('task');
+const deleteIcon = document.getElementsByClassName('delete-icon');
+const toHide = document.getElementsByClassName('to-hide');
 
 const taskList = new ToDo();
 taskList.displayTask();
