+++ src/style.css
@@ -46,6 +46,7 @@ input:focus,
   border-width: 0;
   border: none;
   height: 100%;
+  width: 100%;
 }
 
 #addBtn {
